Work from Monday 12 21 20
I formatted the scatter plots a little more. I think it would be a good idea to settle on one color and one marker. 
I also started looking through the scatter plots and thought they weren't providing the information that we needed so I took them and put them into a box plot comparing them to all the teams over the last 30 years. Having done that, I think it gives a much better picture of teams should look for if they want to be competitive for a WS championship.